services/user_service/requirements.txt
firebase_admin==5.0.3
google-cloud-pubsub